requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
timm>=0.9.0
Pillow>=9.0.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
rouge-score>=0.1.2
sacrebleu>=2.3.0
meteor>=1.0.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
jsonlines>=4.0.0
fairscale>=0.4.0
rad-dino>=0.1.1
